// FrontEnd/src/app/models/TotalHours.ts
export interface TotalHours {
  totalTeacherActivities: number;
  subtotalTeacherActivities: number;
  subtotalInvestigationActivities: number;
  subtotalExtensionActivities: number;
  subtotalAdministrationActivities: number;
  subtotalRepresentationActivities: number;
  subtotalOtherActivities: number;
  totalCai: number;
}